Move cgroup work to dedicated package
No functional changes
