+++ cgroup/cgroup_test.go
@@ -1,4 +1,4 @@
-package systemd
+package cgroup
 
 import "testing"
